Move RensonData and RensonConfigEntry to coordinator module
